Your school: make announcement responsive
The /yourschool page announcement wasn't responsive as the browser was resized, because the window resize event listener wasn't being registered.

The announcement block now has a min-height but can get longer on very narrow displays.  This shouldn't cause a mismatch between the block and the image to its side because the image goes away on narrow displays.