Hour of Code site: no more links to partner sites
This change is intended to remove all links to partner sites such as uk.code.org.

Up until now, the function resolve_url was being called from a lot of content for two purposes: links back to code.org; and links within hourofcode.com.  The former functionality is largely deprecated, since it no longer maps to a partner variant of code.org, but a little code is left in case we missed any links, however it no longer maps to a partner variant.

For content that does link to code.org, it's updated to one of two variants: we include literal references to code.org where the string is user-visible on a page; we changed it to use codeorg_url, a new helper which maps to https://code.org for any link to that site.