Update sync-codeorg-all script to allow for non-interactive execution
Specifically, give it some option parsing that'll let it distinguish
between an interactive mode with `-i` (which works the same as it used
to) and a command-specific mode with `-c` that will run just one of the
steps.

This will eventually be used by the i18n server to break the steps up so
we can run them at different times with lots more flexibility.

As a bonus, we get a nice `--help` output for free:

```bash
$ bin/i18n/sync-codeorg-all.rb --help
  Usage: sync-codeorg-all [options]

  Commands:
    in:    Pull all updated source strings into i18n/locales/sources
    up:    Upload i18n/locales/sources to crowdin
    down:  Download translations from crowdin into i18n/locales
    out:   Distribute translations from i18n/locales out into codebase

  Options:
    -i, --interactive                Run through complete sync interactively
    -c, --command COMMAND            Run a single sync command
    -p, --with-pull-request          Automatically generate pull requests
```