Research: gather data on accessible video player sites
As we pondered which site, or sites, we should use for playing video, we realised we should gather some data on which of them are accessible by students and teachers.

To that end, the signed-in homepage now uses the existing functionality to detect whether a site is reachable (which retrieves an image as a DOM element to bypass CORS restrictions) for four candidate domains.  Those domains are videos.code.org, youtube.com, youtube-nocookie.com, and code.hosted.panopto.com.

The list of accessible sites is reported to Google Analytics as a single string of the form "pn-vc-yc-yt", sorted alphabetically.  Not only does this form of reporting give us data on which sites are accessible, but it tells us which combinations frequently occur together, which seems an interesting datapoint when trying to understand common firewall configurations.