Fix newline crowdin issues
The level data we've been uploading to crowdin has two issues.

First, the yaml system we've been using wraps content like:

    en:
      data:
        instructions:
          2-3 Artist 1 new_instruction: "Hi, I'm an artist. You can write code to make
            me draw almost anything. Use a few blocks to make me draw over the grey lines
            in the picture. (Each line is 100 pixels long)"

This is not a valid way to do multiline strings in yaml. Although there
certainly exist ways to do multiline strings in yaml with the goal of
keeping yaml as a human-editable serialization format, we don't really
care about human-editability for this data and so are now going to
serialize the data like:

    en:
      data:
        instructions:
          2-3 Artist 1 new_instruction: "Hi, I'm an artist. You can write code to make me draw almost anything. Use a few blocks to make me draw over the grey lines in the picture. (Each line is 100 pixels long)"

Second, our data includes carriage returns (`\r`) as well as newlines
(`\n`). This should be fine, but crowdin does not render carriage
returns correctly, instead just rendering them as plaintext. This
results in translators seeing content that looks like

    My snowflake is falling apart!\r
    \r
    Can you figure out what's going wrong?

To fix this, we just strip out all carriage returns, replacing them with
newlines as appropriate