Privacy: teacher sign-up has email preference
Teacher sign-up now has an option to choose email preference.  A choice is not required to submit the form at this time, but not choosing "yes" will default to an email preference opt-in choice of "no".

The option is shown as a drop-down, though a hidden alternative uses radio buttons.  An A/B test is able to show this alternate using something like this:

  $("#optin-container").remove();
  $("#optin2-container").show();