Tests asserting current query count
Extends the existing controller tests for the section_level_progress API to assert the number of queries generated when calling that API in different configurations.  I'm currently exploring possible optimizations to this API, and would like this checked in as a baseline so that future improvements are visible in the diff.