LevelConceptDifficulty keeps attributes on create
Fixes a subtle bug we didn't notice for a long time:

  Given a Level that has no LevelConceptDifficulty
  When a LevelConceptDifficulty is created by passing a `level_concept_difficulty` hash to Level's `assign_attributes` method (via ActiveRecord's [`assign_nested_attributes_for_one_to_one_association`](https://apidock.com/rails/ActiveRecord/NestedAttributes/assign_nested_attributes_for_one_to_one_association) method)
  Then I expect the created LevelConceptDifficulty to have the properties provided in the hash I passed to Level
  But actually all of the concept properties are nil after creation!

The root cause is that `assign_nested_attributes_for_one_to_one_association` calls `assign_attributes` twice when it causes a new LevelConceptDifficulty to be created - once when it initially builds the model instance, and a second time to set the id of the associated Level.  This second call would trigger our custom logic in LevelConceptDifficulty's `assign_attributes` method that unsets any concepts not included in the provided attributes... but since the provided attributes in this second call are only `{level_id: 123}` _all_ concepts would be unset, and we'd end up missing properties.

As a solution, I'm skipping the custom behavior when creating a new LevelConceptDifficulty record, because in that case we shouldn't have to worry about unsetting properties.