Make /learn/local work fully & minor fixes
It turns out that process_forms was writing some information to Solr that it never wrote to MySQL, and once we made learn/local work without talking to Solr, it wasn't able to do proper filtering.  With this change, we write that additional information to MySQL in the forms table's processed_data column, and the query from the browser now looks in the right place for it.

This also fixes a couple small things: the "Review" links on /private/review-class are now visible; and the currently-selected email opt-in is now shown properly when reviewing a submitted class.