Pegasus: improve Solr removal forms backfill script & make it write
The improves the script which backfills pegasus form submissions in ancipation of the Solr removal.

It will no longer write updated timestamps in any of the _at fields, so that timestamps from original processing are all left in place.

It also makes the script write its changes.  A successful read-only rehearsal on production took less than 10 seconds to run and gave the desired output.