Fix performance issue when clearing poste data [skip ui]
Our first attempt to purge a user account on production failed due to a
query timeout.  On further investigation I discovered that the slow
query was [generated by this
line](https://github.com/code-dot-org/code-dot-org/blob/2f1f4618b82d7ae5109f9bd95d95160e2eb5ae24/lib/cdo/delete_accounts_helper.rb#L190)
and was roughly:

```sql
SELECT * FROM pegasus.poste_deliveries WHERE contact_email =
'user@example.com';
```

I EXPLAINed this query and discovered it was causing a table scan on a
30-million-plus row table.  It turns out `poste_deliveries` does not
have a key on `contact_email` but it _does_ have a key on `contact_id`.

I'm switching the account purge logic to do this lookup on `contact_id`.
This is unfortunate though - since we're not actually doing the lookup
on email, it's possible that malformed records in our db will cause us
to leave a matching email intact.  It may be worth adding an email index
to this table so we can be sure we're purging every instance of the
given address from our system.

Note: `poste_deliveries` on production also has a key on `(contact_id,
contact_email)` but that index doesn't show up in my local db, so I'm
assuming it's deprecated.  In any case it didn't prevent a table scan on
the original query because email wasn't listed first.