Broken link checker: write final output before exiting
Our Slack channel was missing the final output from the broken link checker, but looking at the emails sent by crontab containing raw logs, the job was finishing.  Turns out we just need to wait for the final post to Slack to complete before killing the process.

There might be an elegant way to make sure we're waiting for all posts to Slack to complete before exiting, but I'm not sure it'll be quite so concise.