DeleteAccountsHelperTest: Do not count rows
Checking the number of rows in a table before and after a test to ensure our test cleaned up properly _seems_ like a good idea, but introduces flakiness when these tests are run in parallel.  I'm tearing out these checks - I'm now confident that these helpers are working as expected, and we shouldn't be depending on our database state so much.