Remove unneeded options from cron script
So far, any time we need to do this manually I find it easier to open dashboard-console and run ExpiredDeletedAccountPurger myself; so this script really doesn't need to pass through options, it can just depend on the defaults.

I'm also about to make the safety constraint a bit more complex, and don't want to change this layer too.