Volunteer map: don't show filters until location is selected
The volunteer map at https://code.org/volunteer/local initially shows pins within 3000 km of a place near the center of the US, showing a good selection of pins country-wide.  All additional lookups, though, are done within 24km of the specified location.

If the user filtered without specifying a location, it would show pins within 24km of the center, which wasn't really the intent.

The solution is to hide the filters until after a location is chosen.  This way, a location of interest is chosen, and then the filters appear, and then the filtering can begin.