Prefer provider+uid for cache key
Before we used email+param for our partial registration cache key.  This is going to be a problem when we send Clever through the finish-sign-up flow because we don't trust/preserve emails of Clever students and we could get key collisions.

Instead, we're going to use provider+uid+param if possible - this will be unique for any OAuth user during registration.  As a fall-back we're switching to hashed_email+param, to avoid accidentally storing cleartext student emails.

I've also adjusted tests a bit, tearing out some of the cases that were too tightly bound to implementation.  We've still got 100% coverage on this module.  I've also added an assertion that token expiration is preserved.