Log less PII in sign-up tracking [ui skip]
Looks like we were dumping the user object to Redshift with as part of tracking sign-up results.  We do sometimes send PII to Redshift, but dumping the whole user is bad practice, may make for large metrics requests, and in general I don't think we're going to be needing this data.

Instead I'm reducing the user data logged down to id / provider / uid and validation errors.  We can always put fields back in if we need them.