Fix getTime logic
When I added the measures display I modified the logic so that we start counting measures at 1. This change applies that same logic to `getTime("measures")`as well as accounting for any delay between the beginning of a song and the first measure (most noticable right now in Can't Touch This).