hoc2018: fix code.org homepage fade bug on Safari
macOS Safari was exhibiting a bug in which the paint background image faded over the top of the black box & triangle SVG background at first, before they took their rightful place above that image once the fade was complete.  Searching online indicates some known issues with Safari in such a situation when there is a negative z-index involved, as there is here.

This commit adds a workaround that one online commenter suggested.  It seems to have fixed this issue, without affecting rendering quality in Chrome or Safari.