Do not count soft-deleted accounts every day [skip ui]
The query for this metric is timing out (at 30 seconds) almost every day.  Suresh suggested we extend the timeout for this particular query but I hate to keep another long-running query around, even against the reporting DB, if we don't really need it.  This is the least useful monitoring metric being sent by the EDAP and there's probably a better way to get it to our dashboards.  Therefore I'm simply removing it, and when we do need it again (when we actually start "burning down" old inactive accounts) we'll set up something better.