Save when feedback dialog with sharing appears
Fixes https://github.com/code-dot-org/dance-party/issues/384

Instead of performing a save-if-changed when clicking the feedback dialog "Continue" button and skipRunSave is set, we will _always_ perform a fire-and-forget save-if-changed when we are about to show a feedback dialog with a share link.  This is a safe no-op if the project has not changed since the last save (will usually be the case) or if we're not on a project level at all.

This does mean we're performing the save earlier than in the previous case, but this should be fine since it should be impossible to change the project source while the feedback dialog is up.

IMO this is also simpler, with fewer "moving parts" than the previous solution.

Working in this part of the code makes me itch to clean up all of feedback.js

Tests coming momentarily.