Ensure primary_contact_info association is cleared when purging user
The ExpiredDeletedAccountPurger asked for manual intervention this week on an account where the purge was failing with an "email already taken" validation failure, which doesn't make much sense.  Upon further investigation I discovered that the account was multi-auth migrated and had a `primary_contact_info` even though its `authentication_options` collection was empty.

Normally this isn't possible but somehow one account ended up pointing at another account's authentication option:

> Account being purged:
> - id: 2
> - primary_contact_info_id: 10
>
> Other account
> - id: 1
> - primary_contact_info_id: 10
>
> AuthenticationOption
> - id: 10
> - user_id: 1

We should certainly figure out how to avoid this situation in the future, but this change is constrained to making the ExpiredDeletedAccountPurger handle this situation gracefully, by removing the association between the purged account and the AuthenticationOption but _not_ hard-deleting the AuthenticationOption because it's still in use elsewhere.