PL: change approach to moving questions between sections
While the teacher application uses the PAGE_LABELS as helpers to get label strings, the details view used to view the application relies on PAGE_LABELS as the source of truth for which questions to show in which sections and in which order.  As a way to preserve that source of truth, we can borrow a technique in which the client pulls across the label strings from another section.