Remove unused marked setting
This code had formerly been used by
https://github.com/code-dot-org/code-dot-org/pull/9799 to escape HTML
entered in the DSL markdown editor, as part of an effort to discourage
excess custom HTML in those files.

However, https://github.com/code-dot-org/code-dot-org/pull/24506
switched CodeMirror over to using a markdown renderer that doesn't rely
on Marked, at which point this options injection stopped working.
Because we are moving away from marked, and because in the long run we
are going to start restricting the tags we support at the lowest level,
this dead code is simply removed rather than replaced.