Only destroy unassigned peer reviews on unsubmit
The :handle_unsubmit method of UserLevel was being called as a before_save hook, with no conditions.  Within that method we were destroying unassigned PeerReview records on any save, not only when an unsubmit had occurred.

This could cause us to have only one peer review for a `UserLevel`, which in turn could cause our backfill cronjob to create two new reviews for that UserLevel putting it back in a 'pending' state even when it had already been accepted.

The fix is to make the `before_save` hook properly conditional on an unsubmit actually occurring.