Remove unused Honeybadger notification
This query and notification in User#track_level_progress_sync always runs right after PeerReview.create_for_submission, which is guaranteed to create two peer reviews if it creates any at all.  That would explain why we've never run this notification.

If we _do_ fail to create peer reviews, it will be because of an exception causing the wrapper transaction to roll back, and the transaction wrapper would re-raise the exception so it would end up in Honeybadger anyway.